requirements.txt
asgiref==3.8.1
astroid==3.0.2
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.7
dj-database-url==2.2.0
Django==5.0.6
djangorestframework==3.15.1
dnspython==2.6.1
gunicorn==22.0.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
peewee==3.17.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pylint==3.0.3
pymongo==4.6.1
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==2.2.1
wcwidth==0.2.10
whitenoise==6.6.0
